[storybook] DRY out shared datagrid utils between stories
